add fem sandwich solver
