+++ .gitignore
@@ -143,6 +143,9 @@ venv.bak/
 .dmypy.json
 dmypy.json
 
+# Pyright
+pyrightconfig.json
+
 # Pyre type checker
 .pyre/
