+++ pyproject.toml
@@ -22,6 +22,7 @@ numpy = "^1.21.0"
 pandas = "^1.3.0"
 plotly = "^5.0.0"
 matplotlib = "^3.10.9"
+scikit-fem = "^12.0.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.0.0"
@@ -50,7 +51,7 @@ warn_unused_configs = true
 disallow_untyped_defs = true
 
 [[tool.mypy.overrides]]
-module = ["pandas", "pandas.*", "plotly.*"]
+module = ["pandas", "pandas.*", "plotly.*", "skfem", "skfem.*"]
 ignore_missing_imports = true
 
 [tool.pytest.ini_options]
