+++ sandwich_numerical/fem/__init__.py
@@ -1,0 +1,1 @@
+"""Finite-element implementation of the Sandwich numerical method."""
